requirements.txt
python-dotenv==1.0.0
mcp @ git+https://github.com/modelcontextprotocol/python-sdk.git 
ibm_db>=3.2.0 